requirements.txt
<<<<<<< HEAD
Django>=5.2.5
djangorestframework>=3.14
torch>=2.0
ftfy
regex
transformers
clip-anytorch
Pillow
numpy
celery>=5.2
redis>=4.5
=======
Django>=5.2.5
djangorestframework>=3.14
torch>=2.0
ftfy
regex
transformers
clip-anytorch
Pillow
numpy
celery>=5.2
redis>=4.5
>>>>>>> 8b1e1d938e70917f9e7bc0a124a56dd9f9496b7e
